requirements.txt
matplotlib
numpy
pandas
seaborn
keras
scikit-learn
IPython
pyyaml